Update to build otelcol binary, then add to the image
This changes the GHA build approach to execute the builder.sh 'locally' in the pipeline, then make that available to the Dockerfile. The Dockerfile is changed to use the same configuration defaults as the 'vanilla' otelcol-contrib, and checked to make sure it'll boot.

Finally, the otelcol version has been upgraded to 0.91.0.